Extract purchaseItem to listService
